pyproject.toml
[project]
name = "example-schedule-py"
version = "1.0.0"
description = "Periodic function scheduling with Resonate's Python SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "resonate-sdk>=0.6.7",
]
